refactor(rainbot): migrate to worker-shared factories
Rainbot was the only worker that hadn't been refactored to use
@rainbot/worker-shared's createWorkerExpressApp / createWorkerAuthMiddleware /
addWorkerHealthRoutes / startWorkerServer / createJoinHandler /
createLeaveHandler / createVolumeHandler. Pranjeet and Hungerbot already
follow the canonical 5-file pattern (config / state/guild-state / handlers/rpc /
events/voice-state / app + slim index). Rainbot now matches.

Module layout (was: one 780-line index.ts):
- src/config.ts: env vars + logger
- src/state/guild-state.ts: RainbotGuildState, guildStates Map,
  getOrCreateGuildState (with player creation + idle-handler), playback
  helpers (resetPlaybackTiming, markPaused, markResumed, getPlaybackPosition,
  buildPlaybackState, buildQueueState, getStateForRpc), and playNext.
- src/handlers/rpc.ts: createRpcHandlers, which uses the shared
  createJoinHandler/Leave/Volume from worker-shared, with onBeforeLeave
  (clear queue + currentTrack) and onVolumeChange (push to currentResource)
  hooks. Rainbot-specific handlers stay inline (enqueue, skip, pause, stop,
  clear, getQueue, autoplay, replay, seek).
- src/events/voice-state.ts: registerVoiceStateHandlers — rainbot extension
  to the shared auto-follow handler that clears music-specific state when
  the orchestrator leaves.
- src/app.ts: createApp via worker-shared (express + auth + tRPC mount + health + startServer).
- src/index.ts: ~85 lines (was ~780), matches hungerbot's startup shape.

Behavior changes:
- state.volume is now 0-1 throughout (was 0-100). The Raincloud→worker tRPC
  contract was already 0-1 (multiBotService.setVolume normalizes 0-100→0-1
  before calling rainbotClient.volume.mutate), so this aligns rainbot's
  internal representation with the wire format. Fixes a latent bug in
  buildPlaybackState which previously returned volume=100 instead of 1
  (other workers and the StatusResponse path were correct via division).
- handleVolume now only accepts 0-1 (was accept-either). No live caller
  relied on the 0-100 path.
- Latent risk identified in the audit (rainbot's own serverStarted flag
  conflicting with worker-shared's module-level flag) is resolved by
  routing through startWorkerServer.

type-check passes across all 18 tasks; no behavioral test coverage exists
for rainbot worker, so verification is bounded to types + integration via
shared factories that pranjeet/hungerbot already exercise.
